Update scripts path in Dockerfile and mirror.sh
The Docker file and mirror.sh have been updated to make all scripts executable from any part of the application. The path to the scripts has been changed to a '/scripts' directory in both files. This is to improve the flexibility of the codebase. All necessary setup files have been moved to this new directory to maintain the functionality of the code.