Remove test workflow, update README
Removed the test.yml workflow configuration file. This file was unnecessary as we had shifted our testing approach to a different setup not related this particular workflow.

README.md has been updated with more detailed information about features and usage of the 'Repository Mirroring Action'. This update aims to provide clear guidance on how to use the action effectively and acknowledges the original creator of the action.